+++ src/constants.rs
@@ -1,5 +1,5 @@
-pub const WIDTH: usize = 320 as usize;
-pub const HEIGHT: usize = 200 as usize;
+pub const WIDTH: usize = 320;
+pub const HEIGHT: usize = 200;
 pub const GENERATOR_HEIGHT: usize = 2;
 pub const GENERATOR_SIZE: usize = WIDTH * GENERATOR_HEIGHT;
 pub const VISIBLE_HEIGHT: usize = HEIGHT - 5;
